+++ crates/teloxide-core/CHANGELOG.md
@@ -107,6 +107,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deserialization of `ApiError::CantParseEntities` ([#839][pr839])
 - Deserialization of empty (content-less) messages that can sometimes appear as a part of callback query ([#850][pr850], issue [#873][issue873])
 - Type of `chat_id` in `send_game`: `u32` => `ChatId` ([#1066][pr1066])
+- `Requester::Err` was bounded to the `AsResponseParameters`. In case of `RetryAfter(..)` errors the polling is paused for the specified delay instead of falling into the backoff strategy ([#1113][pr1113])
 
 [pr839]: https://github.com/teloxide/teloxide/pull/839
 [pr879]: https://github.com/teloxide/teloxide/pull/879
@@ -116,6 +117,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [pr990]: https://github.com/teloxide/teloxide/pull/990
 [pr990]: https://github.com/teloxide/teloxide/pull/990
 [pr1066]: https://github.com/teloxide/teloxide/pull/1066
+[pr1113]: https://github.com/teloxide/teloxide/pull/1113
 
 ### Changed